refactor: dto 수정
