--- src/main/java/woowacourse/shoppingcart/application/dto/ProductResponse.java
@@ -1,43 +1,0 @@
-package woowacourse.shoppingcart.application.dto;
-
-import woowacourse.shoppingcart.domain.Product;
-
-public class ProductResponse {
-
-    private long id;
-    private String name;
-    private int price;
-    private String imageUrl;
-    private int stock;
-
-    public ProductResponse() {
-    }
-
-    public ProductResponse(Product product) {
-        this.id = product.getId();
-        this.name = product.getName();
-        this.price = product.getPrice();
-        this.imageUrl = product.getImageUrl();
-        this.stock = product.getStock();
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public int getStock() {
-        return stock;
-    }
-}
