+++ src/main/java/woowacourse/shoppingcart/ui/ProductController.java
@@ -1,13 +1,12 @@
 package woowacourse.shoppingcart.ui;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.shoppingcart.application.ProductService;
-import woowacourse.shoppingcart.ui.dto.ProductResponse;
+import woowacourse.shoppingcart.dto.CartItemAddRequest.ProductResponse;
 
 @RestController
 @RequestMapping("/api/products")
@@ -21,10 +20,6 @@ public class ProductController {
 
     @GetMapping
     public ResponseEntity<List<ProductResponse>> products() {
-        List<ProductResponse> products = productService.findAllProducts()
-                .stream()
-                .map(ProductResponse::new)
-                .collect(Collectors.toUnmodifiableList());
-        return ResponseEntity.ok(products);
+        return ResponseEntity.ok(productService.findAllProducts());
     }
 }
